feat: Introduce `wallpapers` module
This module will automatically copy wallpapers into system directory for all image variants.

If image is Gnome-based, it can additionally support advanced features:

- Setting all wallpapers to appear in Gnome settings, Appearance section.
- Including specified light+dark wallpapers
- Setting default wallpaper, including light+dark variant
- Setting various image scaling options, globally &/or per-wallpaper

Gnome additional support, which is not planned or not yet implemented:

- Setting different colors & gradients for background scaling filler (default is black).
- Dynamic wallpapers (can be based on various time conditions)
- ... other things which I don't know or can't remember

Other DEs support:
- Let me know

I tested this thoroughly & it works great!

I made this module to be as easy to maintain as possible in the code. F.e, if some scaling option gets depreciated, it is very easy to remove it in the code, inside `scaling_options` variable.
I also categorized the code & explained as much as possible.

Only thing that can be possibly adjusted in this PR is documentation & maybe some other minor things.