fix(gnome-extensions): Use `wget` instead of `curl`
`wget` might be more reliable for downloading things, which is performed in this module.

Some user faced the error with extension that didn't download fully, causing the `unzip` command to fail.
I tested the same extension & it installed successfully for me.

This was maybe a server error, maybe `curl`'s, but I will still do this change, as I have more trust in `wget` for strictly downloading things.