chore(default-flatpaks): Don't modify the remote title if it's already modified
Wanting to avoid running unnecessary setup in logs as much as possible.

Unfortunately,
`flatpak remote-add` still runs even if repo exists, as it apparently modifies the URL, no matter if it's the same.

I tried to extract & compare input URL & flatpak URL to mitigate this as an condition, but it doesn't work, as flatpak lists repo URL, while we input flatpakref URLs, which are not the same.