feat(platforms): Add support for building all platforms (#472)
* feat(platforms): Add support for building all platforms

This is to support the upcoming multi-platform build feature with the CLI. The build will still work without this, but it will throw up warnings in the log that could throw our users off when trying to diagnose the problem with their image.

This would normally be a very computationally intense process creating all these images. However, since the modules are basically single layer `COPY`'s, this shouldn't take too long.

* Create a central constants file