feat(nushell): Convert the scripts module to nushell (#515)
* feat(nushell): Convert the scripts module to nushell

This follows the new pattern in #503 for trying to be POSIX compatible or use nushell for modules. This does change the scripts module so that the snippets are executed with `sh` instead of `bash`. And it also fixes a bug where a user would create a multi-line snippet only for each line in that one multi-line snippet to be executed separately. Now the entire multi-line snippet entry will execute as one script.

* Update modules/script/script.tsp

Co-authored-by: xyny <60004820+xynydev@users.noreply.github.com>

* Update modules/script/v2/README.md

Co-authored-by: xyny <60004820+xynydev@users.noreply.github.com>

* Update modules/script/v2/README.md

Co-authored-by: xyny <60004820+xynydev@users.noreply.github.com>

* Update modules/script/v2/README.md

Co-authored-by: xyny <60004820+xynydev@users.noreply.github.com>

---------

Co-authored-by: xyny <60004820+xynydev@users.noreply.github.com>